# By Andrey Derkach
# Via Andrey Derkach
* commit '3c198b81dc704e89feef841cc85ec09d1f4adba7 [formerly 14a30838712769c73ac41e86f21f5bd944467949]':
  [DF-8361] Render initialization in the right way
  [DF-8361] fix of x64 warnings
  [DF-8361] fix of x64 warnings


Former-commit-id: 980eef51e73fc7e5d6968079654ea9d948992417

+++ Sources/Internal/Render/Highlevel/StaticOcclusionRenderPass.cpp
@@ -200,7 +200,7 @@ void StaticOcclusionRenderPass::DrawOcclusionFrame(RenderSystem* renderSystem, C
               [](const RenderBatchWithDepthOption& a, const RenderBatchWithDepthOption& b) { return a.first->layerSortingKey < b.first->layerSortingKey; });
 
     target.blockIndex = blockIndex;
-    target.queryBuffer = rhi::CreateQueryBuffer(meshBatchesWithDepthWriteOption.size());
+    target.queryBuffer = rhi::CreateQueryBuffer(static_cast<uint32>(meshBatchesWithDepthWriteOption.size()));
     target.frameRequests.resize(meshBatchesWithDepthWriteOption.size(), nullptr);
 
     passConfig.queryBuffer = target.queryBuffer;
